move OperatorPrecedence from linter to ruff_python_ast
